requirements.txt
apache-airflow==2.5.0
numpy==1.24.1
pandas==1.5.3
pyyaml==6.0.1
scikit-learn==1.6.1
